requirements.txt
langchain-google-genai
langchain
langchain-core
streamlit
python-dotenv
langchain_community
langchain_groq
langserve
fastapi
uvicorn 
sse_starlette
